% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fromBase}
\alias{fromBase}
\title{Convert base representation of a number into integer}
\usage{
fromBase(x, base = 2L)
}
\arguments{
\item{x}{vector of positive integer coefficients representing the number
in base that is \code{base}.}

\item{base}{positive integer representing the base.}
}
\value{
The function returns a positive integer that is a
conversion from \code{base} under given coefficients \code{x}.
}
\description{
Converts base representation of a number into integer.
}
\examples{
fromBase(c(1, 2, 0, 2, 3), 5)
}
