% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seqPrimes}
\alias{seqPrimes}
\title{Sequence of prime numbers}
\usage{
seqPrimes(n)
}
\arguments{
\item{n}{positive integer representing the number of sequence elements.}
}
\value{
The function returns a numeric vector containing 
first \code{n} prime numbers. The current (naive) implementation of the 
algorithm is not efficient in terms of speed so it is suited for low 
\code{n < 10000} but requires just O(n) memory usage.
}
\description{
Calculates the sequence of prime numbers.
}
\examples{
seqPrimes(10)
}
