% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{toBase}
\alias{toBase}
\title{Convert integer value to other base}
\usage{
toBase(x, base = 2L)
}
\arguments{
\item{x}{positive integer representing the number to convert.}

\item{base}{positive integer representing the base.}
}
\value{
The function returns a numeric vector containing 
representation of \code{x} in a base given in \code{base}.
}
\description{
Converts integer value to other base.
}
\examples{
toBase(888, 5)
}
